backend/requirements.txt
# Django and REST Framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database
psycopg2-binary>=2.9.0

# ML/AI Libraries (using latest compatible versions)
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.10.0
huggingface-hub>=0.16.0

# Image Processing
pdf2image>=1.16.0
Pillow>=10.0.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Utils
python-decouple>=3.8
python-multipart>=0.0.6

# Development
black>=23.0.0
flake8>=6.0.0